Fixes the 'Display not ready' error and album playing
Right now, if I do a search and press `RET` on a track it says `[Display not
ready]` on the minibufer and it does nothing (I think it has to do with not
having any `action` defined in helm, just `action-transformer`).
With this fix, pressing `RET` or `TAB` on the track selected will trigger the
playing of the track.

Also, the `Play album` wasn't behaving properly. It just played the first song
of the track and then it stopped. The fix there has to do with the URI that is
passed to Spotify. And I think it isn't necesary to make another http call to
get the album information, its's just there in the first request, so i wipe out
that function (`spotify-get-track`) entirely